adding confirm to delete watch listing

+++ src/components/YourWatches.jsx
@@ -12,12 +12,15 @@ function YourWatches (props) {
 
     const watch = props.watch
 
-    const deleteUser = async () => {
+    const deleteWatch = async () => {
         const userDoc = doc(db, 'watches', watch.id)
         const imageRef = ref(storage, watch.imageUrl)
-        await deleteDoc(userDoc)
-        await deleteObject(imageRef)
-        navigate('/')
+        const confirmed = window.confirm("Are you sure you want to delete this listing")
+            if(confirmed) {
+                await deleteDoc(userDoc)
+                await deleteObject(imageRef)
+                navigate('/')
+            }
     }
 
     return (
@@ -28,7 +31,7 @@ function YourWatches (props) {
                 <div className={styles.watchCardDetails}>
                     <h2 className={styles.watchCardName}>{watch.name}</h2>
                     <p className={styles.watchCardPrice}>${watch.price}</p>
-                    <button className={styles.deleteButton} onClick={() => deleteUser()}>Delete Listing</button>
+                    <button className={styles.deleteButton} onClick={() => deleteWatch()}>Delete Listing</button>
                 </div>
             </div>
     )
